backend/src/requirements.txt
alembic==1.8.1
aniso8601==6.0.0
Babel==2.10.3
Click==7.0
distlib==0.3.5
ecdsa==0.18.0
filelock==3.8.0
Flask==1.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==1.0.1
future==0.18.2
importlib-metadata==4.12.0
itsdangerous==1.1.0
Jinja2==2.10.1
jose==1.0.0
Mako==1.2.1
MarkupSafe==1.1.1
packaging==21.3
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycryptodome==3.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2019.1
rsa==4.9
six==1.12.0
SQLAlchemy==1.3.4
virtualenv==20.16.3
Werkzeug==0.15.5
wrapt==1.11.1
WTForms==3.0.1
zipp==3.8.1
